requirements.txt
requests==2.31.0
selenium==4.15.2
pandas==2.1.4
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0 